@@ -1,5 +1,8 @@
-﻿namespace Facteur.Tests
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace Facteur.Tests
 {
+    [ExcludeFromCodeCoverage]
     public class TestMailModel
     {
         public string Name { get; set; }
