@@ -1,8 +1,11 @@
-﻿using System.Diagnostics.CodeAnalysis;
+﻿using System;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Net.Http.Headers;
+using System.Threading.Tasks;
 using Microsoft.Graph;
 using Microsoft.Identity.Client;
+
 using t = System.Threading.Tasks;
 
 namespace Facteur.MsGraph
@@ -13,13 +16,16 @@ namespace Facteur.MsGraph
     [ExcludeFromCodeCoverage]
     public class GraphMailer : IMailer
     {
+        private readonly IEmailComposer _composer;
+
         /// <summary>
         /// Initializes a new instance of the <see cref="GraphMailer"/> class
         /// </summary>
         /// <param name="credentials"></param>
-        public GraphMailer(GraphCredentials credentials)
+        public GraphMailer(GraphCredentials credentials, IEmailComposer composer = null)
         {
             Credentials = credentials;
+            _composer = composer ?? new EmailComposer();
         }
 
         private GraphCredentials Credentials { get; }
@@ -30,30 +36,6 @@ namespace Facteur.MsGraph
         /// <param name="request">The request</param>
         /// <returns></returns>
         public async t.Task SendMailAsync(EmailRequest request)
-        {
-            GraphServiceClient graphClient = await ConnectClient().ConfigureAwait(false);
-            Message message = new()
-            {
-                Subject = request.Subject,
-                Body = new ItemBody { ContentType = BodyType.Html, Content = request.Body },
-                ToRecipients = request.To.Select(x => new Recipient { EmailAddress = new EmailAddress { Address = x } }),
-                Attachments = request.AddAttachments()
-            };
-
-            await graphClient.Users[Credentials.From]
-                .SendMail(message, false)
-                .Request()
-                .PostAsync()
-                .ConfigureAwait(false);
-        }
-
-        /// <summary>
-        /// Sends the mail.
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="request">The request.</param>
-        /// <returns></returns>
-        public async t.Task SendMailAsync<T>(EmailRequest<T> request) where T : class
         {
             GraphServiceClient graphClient = await ConnectClient().ConfigureAwait(false);
             Message message = new()
@@ -72,6 +54,9 @@ namespace Facteur.MsGraph
                 .PostAsync()
                 .ConfigureAwait(false);
         }
+
+        public async Task SendMailAsync(Func<IEmailComposer, Task<EmailRequest>> compose)
+            => await SendMailAsync(await compose(_composer));
 
         protected async t.Task<GraphServiceClient> ConnectClient()
         {
