@@ -1,5 +1,8 @@
-﻿namespace Facteur.Resolvers.ViewModel.Tests
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace Facteur.Resolvers.ViewModel.Tests
 {
+    [ExcludeFromCodeCoverage]
     public class BaseModel
     {
         public string Name { get; set; }
@@ -7,10 +10,12 @@
         public string Email { get; set; }
     }
 
+    [ExcludeFromCodeCoverage]
     public class TestMailModel : BaseModel
     {
     }
 
+    [ExcludeFromCodeCoverage]
     public class TestViewModel : BaseModel
     {
     }
