@@ -1,4 +1,4 @@
-using System;
+using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -6,12 +6,13 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 namespace Facteur.TemplateProviders.IO.Tests
 {
     [TestClass]
+    [ExcludeFromCodeCoverage]
     public class AppDirectoryTemplateProviderTests
     {
         [TestMethod]
         public async Task AppDirectoryTemplateProvider_HasCorrectData_ShouldReturnPath()
         {
-            ITemplateProvider provider = new AppDirectoryTemplateProvider("Templates", ".dhtml");
+            ITemplateProvider provider = new AppDirectoryTemplateProvider("Templates", ".sbnhtml");
             string template = await provider.GetTemplate("Test");
 
             Assert.IsFalse(string.IsNullOrEmpty(template));
