@@ -1,25 +1,26 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 
 namespace Facteur
 {
     /// <summary>
-    /// Represents an e-mail service
+    /// An email service that can send out emails.
     /// </summary>
     public interface IMailer
     {
         /// <summary>
-        /// Sends out an e-mail
+        /// Sends out an email
         /// </summary>
         /// <param name="request">The request</param>
         /// <returns>An instance of <see cref="System.Threading.Tasks.Task"/></returns>
         Task SendMailAsync(EmailRequest request);
 
         /// <summary>
-        /// Sends out an e-mail
+        /// Sends out an email
         /// </summary>
         /// <typeparam name="T">The model type</typeparam>
         /// <param name="request">The request</param>
         /// <returns>An instance of <see cref="System.Threading.Tasks.Task"/></returns>
-        Task SendMailAsync<T>(EmailRequest<T> request) where T : class;
+        Task SendMailAsync(Func<IEmailComposer, Task<EmailRequest>> compose);
     }
 }
