@@ -3,15 +3,15 @@
 namespace Facteur
 {
     /// <summary>
-    /// 
+    /// Represents a type that can retrieve email templates.
     /// </summary>
     public interface ITemplateProvider
     {
         /// <summary>
-        ///
+        /// Gets the email template
         /// </summary>
-        /// <param name="name"></param>
-        /// <returns></returns>
+        /// <param name="name">The template's name</param>
+        /// <returns>The template contents</returns>
         Task<string> GetTemplate(string name);
     }
 }
